refactor: extract get documents and insurance related resources
